docs: nest component recipes under components
- move individual implementation examples under related components
- label them as such
- note: some recipes are used in stories, which breaks convention
- update snapshots
- remove plop integrations for creating new recipes and pages
- remove lingering recipes
