--- .storybook/recipes/ButtonActionCalloutCard/index.ts
@@ -1,1 +1,0 @@
-export { ButtonActionCalloutCard as default } from './ButtonActionCalloutCard';
