--- .storybook/recipes/CardWithNotification/index.ts
@@ -1,1 +1,0 @@
-export { CardWithNotification as default } from './CardWithNotification';
