--- .storybook/recipes/Filters/index.ts
@@ -1,1 +1,0 @@
-export { Filters as default } from './Filters';
