--- .storybook/recipes/NumberIconList/NumberIconList.module.css
@@ -1,14 +1,0 @@
-@import '../../../src/design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # NUMBER ICON LIST
-\*------------------------------------*/
-
-/**
- * Container recipe for the number icons to flex wrap them.
- */
-.number-icon-list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--eds-size-half);
-}
