--- .storybook/recipes/NumberIconList/index.ts
@@ -1,1 +1,0 @@
-export { NumberIconList as default } from './NumberIconList';
