--- .storybook/recipes/StackedCardsToTable/index.ts
@@ -1,1 +1,0 @@
-export { StackedCardsToTable as default } from './StackedCardsToTable';
