+++ package.json
@@ -63,8 +63,6 @@
     "storybook:axe": "yarn run build:storybook && yarn run storybook:axeOnly",
     "storybook:axeOnly": "axe-storybook --build-dir storybook-static",
     "plop": "plop component",
-    "plop:page": "plop page",
-    "plop:recipe": "plop recipe",
     "test": "jest",
     "test:ci": "yarn run test --ci --coverage && cat ./coverage/lcov.info | codecov",
     "types": "tsc --noEmit"
