--- plop-templates/Page/Page.stories.tsx.hbs
@@ -1,17 +1,0 @@
-import React from 'react';
-import { StoryObj, Meta } from '@storybook/react';
-
-import { {{pascalCase name}} } from './{{pascalCase name}}';
-
-export default {
-  title: 'Pages/{{pascalCase name}}',
-  component: {{pascalCase name}},
-} as Meta<Args>;
-
-type Args = React.ComponentProps<typeof {{pascalCase name}}>;
-
-export const Default: StoryObj<Args> = {
-  args: {
-
-  },
-};
