--- plop-templates/Page/Page.tsx.hbs
@@ -1,14 +1,0 @@
-import React from 'react';
-
-import {
-  PageHeader,
-} from '../../../src';
-
-import { PageShell } from '../../recipes/PageShell/PageShell';
-
-export const {{pascalCase name}} = () => (
-  <PageShell>
-    <PageHeader title="Page title" />
-  </PageShell>
-);
-
