--- plop-templates/Page/index.ts.hbs
@@ -1,1 +1,0 @@
-export { {{pascalCase name}} as default } from './{{pascalCase name}}';
