--- plop-templates/Recipe/Recipe.module.css.hbs
@@ -1,12 +1,0 @@
-@import '../../../src/design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # {{spacedUpperCase name}}
-\*------------------------------------*/
-
-/**
- * {{pascalCase name}}
- */
-.{{dashCase name}} {
-  
-}
