--- plop-templates/Recipe/Recipe.tsx.hbs
@@ -1,28 +1,0 @@
-import clsx from 'clsx';
-import React from 'react';
-import styles from './{{pascalCase name}}.module.css';
-
-export interface Props {
-  /**
-   * CSS class names that can be appended to the component.
-   */
-  className?: string;
-}
-
-/**
- * TODO: update this comment with a description of the recipe.
- */
-export const {{pascalCase name}} = ({
-  className,
-  ...other
-}: Props) => {
-  const componentClassName = clsx(styles['{{dashCase name}}'], className);
-  return (
-    <div
-      className={componentClassName}
-      {...other}
-    >
-    Hello!
-    </div>
-  );
-};
