+++ plopfile.js
@@ -86,63 +86,4 @@ module.exports = (plop) => {
       },
     ],
   });
-  plop.setGenerator('recipe', {
-    description: 'Create a recipe',
-    prompts: [
-      {
-        type: 'input',
-        name: 'name',
-        message: 'What is your recipe name (e.g. ComponentName)?',
-      },
-    ],
-    actions: [
-      {
-        type: 'add',
-        path: '.storybook/recipes/{{pascalCase name}}/{{pascalCase name}}.tsx',
-        templateFile: 'plop-templates/Recipe/Recipe.tsx.hbs',
-      },
-      {
-        type: 'add',
-        path: '.storybook/recipes/{{pascalCase name}}/{{pascalCase name}}.stories.tsx',
-        templateFile: 'plop-templates/Recipe/Recipe.stories.tsx.hbs',
-      },
-      {
-        type: 'add',
-        path: '.storybook/recipes/{{pascalCase name}}/{{pascalCase name}}.module.css',
-        templateFile: 'plop-templates/Recipe/Recipe.module.css.hbs',
-      },
-      {
-        type: 'add',
-        path: '.storybook/recipes/{{pascalCase name}}/index.ts',
-        templateFile: 'plop-templates/Recipe/index.ts.hbs',
-      },
-    ],
-  });
-  plop.setGenerator('page', {
-    description: 'Create a page',
-    prompts: [
-      {
-        type: 'input',
-        name: 'name',
-        message: 'What is your page name (e.g. Homepage)?',
-      },
-    ],
-    actions: [
-      {
-        type: 'add',
-        path: '.storybook/pages/{{pascalCase name}}/{{pascalCase name}}.tsx',
-        templateFile: 'plop-templates/Page/Page.tsx.hbs',
-      },
-      {
-        type: 'add',
-        path: '.storybook/pages/{{pascalCase name}}/{{pascalCase name}}.stories.tsx',
-        templateFile: 'plop-templates/Page/Page.stories.tsx.hbs',
-      },
-      {
-        type: 'add',
-        path: '.storybook/pages/{{pascalCase name}}/index.ts',
-        templateFile: 'plop-templates/Page/index.ts.hbs',
-      },
-    ],
-  });
 };
