+++ src/components/Menu/Menu.test.tsx
@@ -9,8 +9,11 @@ import * as stories from './Menu.stories';
 
 const { Default } = composeStories(stories);
 
+// Remove the tests with `play` in, b/c it causes the test runner to be unhappy
+const { Opened, ...staticStories } = stories;
+
 describe('<Menu />', () => {
-  generateSnapshots(stories, {
+  generateSnapshots(staticStories, {
     getElement: async () => {
       const user = userEvent.setup();
       const triggerButton = await screen.findByRole('button');
