+++ src/components/NumberIcon/NumberIcon.stories.tsx
@@ -13,6 +13,7 @@ export default {
     'aria-label': 'Step 1',
     number: 1,
   },
+  decorators: [(Story) => <div className="p-8">{Story()}</div>],
 } as Meta<Args>;
 
 type Args = React.ComponentProps<typeof NumberIcon>;
@@ -75,6 +76,46 @@ export const DifferentNumbers: StoryObj<Args> = {
       <NumberIcon {...args} aria-label="Step 76" number={76} />
       <NumberIcon {...args} aria-label="Step 87" number={87} />
       <NumberIcon {...args} aria-label="Step 98" number={98} />
+    </div>
+  ),
+};
+
+export const NumberIconList: StoryObj<Args> = {
+  parameters: {
+    badges: ['1.0', 'implementationExample'],
+  },
+  render: () => (
+    <div className="flex flex-wrap gap-1">
+      <NumberIcon aria-label="Item 1" number={1} size="sm" />
+      <NumberIcon
+        aria-label="Item 2"
+        incomplete
+        number={2}
+        numberIconTitle="incomplete step 2"
+        size="sm"
+      />
+      <NumberIcon aria-label="Item 3" number={3} size="sm" />
+      <NumberIcon
+        aria-label="Item 4"
+        incomplete
+        number={4}
+        numberIconTitle="incomplete step 4"
+        size="sm"
+      />
+      <NumberIcon
+        aria-label="Item 5"
+        incomplete
+        number={5}
+        numberIconTitle="incomplete step 5"
+        size="sm"
+      />
+      <NumberIcon
+        aria-label="Item 6"
+        incomplete
+        number={6}
+        numberIconTitle="incomplete step 6"
+        size="sm"
+      />
     </div>
   ),
 };
