+++ src/components/Slider/Slider.tsx
@@ -84,7 +84,7 @@ export type Props = React.InputHTMLAttributes<HTMLInputElement> & {
  *
  * Allows input of a value via dragging a thumb along a track.
  * Strict: This slider requires a visual indicator of value/markers.
- * Please check out our recipes for possible ideas.
+ * Please check out our examples for possible ideas.
  *
  * NOTE: This component requires `label` or `aria-label` prop
  */
