+++ src/components/Table/StackedCardsToTable.tsx
